#12 Create a BucketInfo class to encapsulate the info about the parent bucket that is about to be split. Had to move the log2 and entropy functions back to FeatureUtil so it could be shared.
