+++ src/main/scala/org/apache/spark/mllib/feature/ThresholdFinder.scala
@@ -18,30 +18,21 @@
 package org.apache.spark.mllib.feature
 
 import org.apache.spark.Logging
+import FeatureUtils._
+
 
 /**
   * Base trait for threshold finders.
   */
 trait ThresholdFinder extends Serializable with Logging {
 
-  private val LOG2 = math.log(2)
-
-  /** @return log base 2 of x */
-  val log2 = { x: Double => math.log(x) / LOG2 }
-
   /**
-    * @param freqs sequence of integer frequencies.
-    * @param n the sum of all the frequencies in the list.
-    * @return the total entropy
+    * @param bucketInfo info about the parent bucket
+    * @param leftFreqs frequencies to the left
+    * @param rightFreqs frequencies to the right
+    * @return the MDLP criterion value, and the weighted entropy value
     */
-  def entropy(freqs: Seq[Long], n: Long) = {
-    -freqs.aggregate(0.0)(
-      { case (h, q) => h + (if (q == 0) 0 else (q.toDouble / n) * log2(q.toDouble / n))},
-      { case (h1, h2) => h1 + h2 }
-    )
-  }
-
-  def calcCriterionValue(s: Double, hs: Double, k: Long,
+  def calcCriterionValue(bucketInfo: BucketInfo,
                          leftFreqs: Seq[Long], rightFreqs: Seq[Long]): (Double, Double) = {
     val k1 = leftFreqs.count(_ != 0)
     val s1 = if (k1 > 0) leftFreqs.sum else 0
@@ -49,10 +40,10 @@ trait ThresholdFinder extends Serializable with Logging {
     val k2 = rightFreqs.count(_ != 0)
     val s2 = if (k2 > 0) rightFreqs.sum else 0
     val hs2 = entropy(rightFreqs, s2)
-    val weightedHs = (s1 * hs1 + s2 * hs2) / s
-    val gain = hs - weightedHs
-    val delta = log2(math.pow(3, k) - 2) - (k * hs - k1 * hs1 - k2 * hs2)
-    (gain - (log2(s - 1) + delta) / s, weightedHs)
+    val weightedHs = (s1 * hs1 + s2 * hs2) / bucketInfo.s
+    val gain = bucketInfo.hs - weightedHs
+    val delta = log2(math.pow(3, bucketInfo.k) - 2) - (bucketInfo.k * bucketInfo.hs - k1 * hs1 - k2 * hs2)
+    (gain - (log2(bucketInfo.s - 1) + delta) / bucketInfo.s, weightedHs)
   }
 
 }
